django/requirements.txt
# The Django Framework
Django==4.1.7
# Security for communication with the react frontend
django-cors-headers==4.2.0
# API library
djangorestframework==3.14.0
# User authentication using tokens
djangorestframework-simplejwt==5.3.0
# Gunicorn HTTP server
gunicorn==21.2.0
# PostgreSQL to Python library
psycopg2==2.9.9
# Environment Variables
django-environ==0.11.2
# Async and scheduled tasks
celery==5.3.6
# celery task scheduler helper
django-celery-beat==2.6.0
# celery save results with django
django-celery-results==2.5.1
# HTTP requests from python
requests==2.28.2
# Filtering querysets, used in views
django-filter=24.2